agentsleague-maf/requirements.txt
# ── Microsoft Agent Framework (pin exact version — not GA yet) ────────────
agent-framework-core==1.0.0b260212
agent-framework-azure-ai==1.0.0b260212
agent-framework-orchestrations==1.0.0b260212
agent-framework-devui==1.0.0b260212
agent-framework-chatkit==1.0.0b260212

# ── Azure SDKs ─────────────────────────────────────────────────────────────
azure-ai-projects>=2.0.0b3          # Foundry Agent Service SDK (Tier 1 profiler)
azure-identity>=1.25.0              # DefaultAzureCredential for Foundry + OTEL
azure-storage-blob>=12.0.0          # Transitive dep of azure-ai-projects
azure-monitor-opentelemetry         # App Insights OTEL exporter
opentelemetry-semantic-conventions-ai==0.4.13

# ── LLM / OpenAI ──────────────────────────────────────────────────────────
openai>=2.0.0                       # Azure OpenAI GPT-4o (Tier 2 profiler fallback)

# ── UI ────────────────────────────────────────────────────────────────────
streamlit>=1.54.0
plotly>=6.0.0

# ── Data validation ───────────────────────────────────────────────────────
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── Data / reporting ──────────────────────────────────────────────────────
pandas>=2.0.0
reportlab>=4.0.0

# ── Evaluation (LLM-as-judge) ─────────────────────────────────────────────
# Uncomment to activate CoherenceEvaluator, RelevanceEvaluator, FluencyEvaluator
# azure-ai-evaluation>=1.0.0

# ── Dev / test ────────────────────────────────────────────────────────────
pytest>=9.0.0
pytest-asyncio>=0.23.0
rich>=13.7.0
